pyproject.toml
[tool.poetry]
name = "yanfastapiadmin"
version = "0.1.0"
description = ""
authors = ["blowyourbra1n <blowyourbra1n@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqladmin = "^0.16.0"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
itsdangerous = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
